Move SnapshotLifecycleStats
to o.e.xpack.core.slm, so it will match SnapshotLifecycleMetadata and
SnapshotLifecyclePolicy
